% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_of_kernel.R
\name{choice_of_kernel}
\alias{choice_of_kernel}
\title{Defines the desired kernel for our problem.}
\usage{
choice_of_kernel(x, arg = NULL)
}
\arguments{
\item{x}{A    number/row vector.}

\item{arg}{The optional kernel (Alphabetical)}
}
\value{
The	kernelized values.
}
\description{
Defines the desired kernel for our problem.
}
\note{
If argument is not alphabetical, it returns the Gaussian Kernel
}

