% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dParzenRosenblatt.R
\name{dparzen_rosenblatt}
\alias{dparzen_rosenblatt}
\title{The first and second derivative of a density based on the Parzen-Rosenblatt density estimator}
\usage{
dparzen_rosenblatt(x, h, d = 1)
}
\arguments{
\item{x}{The observations}

\item{h}{The bandwidth}

\item{d}{The first or second derivative}
}
\value{
The estimated f^(r)(x) for every observation
}
\description{
The first and second derivative of a density based on the Parzen-Rosenblatt density estimator
}
\note{
By default we use the Gaussian Kernel
}

