% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnonparpoly.R
\name{gnonparpoly}
\alias{gnonparpoly}
\title{The general Non parametric regression estimate}
\usage{
gnonparpoly(y, x, h = "Silverman", p = 0, distance = "Eucledean",
  kernel = "Gaussian")
}
\arguments{
\item{y}{The regressant}

\item{x}{The observations}

\item{h}{The bandwidth}

\item{p}{The order of the local polynomial (by default p=0 Nadaraya Watson)}

\item{distance}{The distance metric}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The estimated f for every observation
}
\description{
The general Non parametric regression estimate
}

