% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNgradient.R
\name{gradient}
\alias{gradient}
\title{The gradient of the three activation functions used in neural networks (sigmoidal,tanh and linear)}
\usage{
gradient(x, arg = "tanh")
}
\arguments{
\item{x}{The inputs.}

\item{arg}{The option of one of the three training functions (tanh by default)}
}
\value{
The	gradient of the activation function.
}
\description{
The gradient of the three activation functions used in neural networks (sigmoidal,tanh and linear)
}

