% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernpredictor.R
\name{kernpredictor}
\alias{kernpredictor}
\title{Non parametric predictor}
\usage{
kernpredictor(my_x, y, x, h = 0.5, p = 0, distance = "Eucledean",
  kernel = "Gaussian")
}
\arguments{
\item{my_x}{The row vector of values (or scalar value) wrt which we do prediction}

\item{y}{The given values before the prediction}

\item{x}{The previous observations of x}

\item{h}{The bandwidth}

\item{p}{The order of the local polynomial (by default p=0)}

\item{distance}{The distance metric}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The predicted y corresponding to my_x
}
\description{
Non parametric predictor
}

