% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCnpreg.R
\name{MCnpreg}
\alias{MCnpreg}
\title{A Monte Carlo simulation}
\usage{
MCnpreg(true, error, x, testh = FALSE, p = 0, method = "Gaussian")
}
\arguments{
\item{true}{The true model}

\item{error}{The error component of the data generating function}

\item{x}{The regressor(s)}

\item{testh}{If we want to test the evolution of the CV criterion with different smoothing parameters (False by default)}

\item{p}{The order of the local polynomial fit (local constant by default)}

\item{method}{The kernel function (Gaussian by default)}
}
\description{
A Monte Carlo simulation
}

