% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinonparpoly1.R
\name{multinonparpoly}
\alias{multinonparpoly}
\title{The general Non parametric regression estimate}
\usage{
multinonparpoly(y, x, h = "Silverman", p = 1, kernel = "Gaussian")
}
\arguments{
\item{y}{The regressant}

\item{x}{The observations}

\item{h}{The bandwidth}

\item{p}{The order of the polynomial (either 0 or 1)}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The estimated f for every observation
}
\description{
The general Non parametric regression estimate
}

