% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNtraining.R
\name{NNtrainQL}
\alias{NNtrainQL}
\title{A two layer neural network for univariate regression (one input for one output) with quadratic loss.}
\usage{
NNtrainQL(y, x, iter = 200, hidden = 3, eta = 0.005, ahidden = "tanh",
  aout = "Linear")
}
\arguments{
\item{x}{The inputs.}

\item{iter}{The number of desired iterations.
@param hidden The number of hidden units in the hidden layer.
@param eta The speed of learning (0.005 by default).
@param ahidden The option of the activation function for the hidden layer (tanh by default).}

\item{aout}{The option of the activation function for the output layer (linear by default).}
}
\value{
The output after training and the MSE per iteration (as a list).
}
\description{
@param y The target output.
}

