% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonParPolyLSCV.R
\name{nonparpolyLSCV}
\alias{nonparpolyLSCV}
\title{The Nonparametric polynomial regression with the Least Squares Cross Validation optimal bandwidth}
\usage{
nonparpolyLSCV(y, x, h = 0.5, method = "Gaussian", p = 0)
}
\arguments{
\item{y}{The regressant}

\item{x}{The regressors}

\item{h}{The bandwidth}

\item{method}{The desired kernel function}

\item{p}{Order of polynomial local fit}
}
\value{
The vector of fitted values produced by the Leave one out CV
}
\description{
The Nonparametric polynomial regression with the Least Squares Cross Validation optimal bandwidth
}

