% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptBandwidth.R
\name{the_bandwidth}
\alias{the_bandwidth}
\title{Automated Bandwidth Selection}
\usage{
the_bandwidth(x, method = "Silverman", y = NULL, p = 1, k = NULL)
}
\arguments{
\item{x}{The observations}

\item{method}{The desired method of selection (Silverman's Rule of Thumb by default, if numerical then this is the user defined value of h)}

\item{y}{If we need the cross validation criterion the regressant}

\item{p}{For the nonparpoly function (order of polynomial local fit)}

\item{k}{If we need the k-nearest neighborhood}
}
\value{
The resulting bandwidth (either scalar or vector)
}
\description{
Automated Bandwidth Selection
}

